+++ CHANGES.rst
@@ -8,6 +8,8 @@ Changelog
 .. note:: This version is not yet released and is under active development.
 
 * [snap]: Add support for ``snap`` on Linux.
+* [mpm]: Add ``--time`` / ``--no-time`` flag to show elapsed execution time.
+  Closes #9.
 * [mpm]: Replace internal helpers with upstreamed ``boltons`` utils.
 * [mpm]: Allow for manager-specific search path to help hunting down CLIs.
 * [mpm]: Bump requirement to ``boltons 20.1.1``.